python-publisher/requirements.txt
paho-mqtt==1.6.1
numpy==1.24.3
pandas==2.0.3
